Separate messages from sorted messages with dates
This change was made to make sure that the array of message objects is
not modified directly, in order to avoid changing data unnecessarily.
The idea is that the sorted messages are used to produce the view,
while the message array just keeps the data that is produced by the
user and the chatbots.

One exemple where it could be a problem is when retrieving messages from a
database. If you wanted to synchronize the local record
with the db, it would be better not to have modified the array that
contains the local record. This way you make sure data flows only one way.